Solved Task 7 and 8

+++ 24.08.2018(Vacation)/methods/Methods_Task_5/src/Student.java
@@ -31,7 +31,7 @@ public class Student {
         this.points = points;
     }
     public void setgradePointAverage (double gradePointAverage){
-        this.gradePointAverage = creditHours/points;
+        this.gradePointAverage = (double) creditHours/ (double) points;
     }
     public int getidNUmber(){
         return idNUmber;
